aarch64: Remove unnecessary asm impl for BSD
The implementations of SafeFetch32 and SafeFetchN in bsd_aarch64.S was
actually x86_64 asm, and not included in the build. So it should be safe
to remove.

The implementation of aarch64_get_thread_helper in
threadLS_bsd_aarch64.S was a duplicate of the implementation for Linux,
that was meant to reduce the number of registers that needed to be saved
on the stack before calling Thread::current_thread(). However, the call
site in macroAssembler_aarch64.cpp saved all the registers anyways for
all non-Linux platforms, so I don't think there's anything to save on
this implementation.

This patch switches to using the C++ implementation of the helper for
all BSD platforms. The same as for OSX and OpenBSD anyways.

This work is sponsored by The FreeBSD Foundation