Let hotspot agent on BSD/Aarch64 lookup threads
This enables getting thread info from attached VMs via the
serviceability API on Aarch64 BSD implementations. This again enables
related features of jhsdb, clhsdb and hsdb.

The actual implementation is shared with MacOSX. Since the underlying
Thread implementation is shared the same way, this should be fine. All
this patch does is enabling the feature for the bsd/aarch64 combo.

This work was sponsored by: The FreeBSD Foundation